+++ crypto/include/aes_icm_ext.h
@@ -57,7 +57,7 @@
 typedef struct {
     v128_t counter; /* holds the counter value          */
     v128_t offset;  /* initial offset value             */
-    int key_size;
+    size_t key_size;
     EVP_CIPHER_CTX *ctx;
 } srtp_aes_icm_ctx_t;
 
@@ -71,7 +71,7 @@ typedef struct {
     v128_t offset;  /* initial offset value             */
     v128_t stream_block;
     size_t nc_off;
-    int key_size;
+    size_t key_size;
     mbedtls_aes_context *ctx;
 } srtp_aes_icm_ctx_t;
 
@@ -87,7 +87,7 @@ typedef struct {
 typedef struct {
     v128_t counter;
     v128_t offset;
-    int key_size;
+    size_t key_size;
     uint8_t iv[16];
     NSSInitContext *nss;
     PK11SymKey *key;
