+++ test/util.c
@@ -115,11 +115,11 @@ uint8_t nibble_to_hex_char(uint8_t nibble)
  * hex_string_to_octet_string converts a hexadecimal string
  * of length 2 * len to a raw octet string of length len
  */
-int hex_string_to_octet_string(char *raw, char *hex, int len)
+size_t hex_string_to_octet_string(char *raw, char *hex, size_t len)
 {
     uint8_t x;
     int tmp;
-    int hex_len;
+    size_t hex_len;
 
     hex_len = 0;
     while (hex_len < len) {
@@ -141,10 +141,10 @@ int hex_string_to_octet_string(char *raw, char *hex, int len)
     return hex_len;
 }
 
-char *octet_string_hex_string(const void *s, int length)
+char *octet_string_hex_string(const void *s, size_t length)
 {
     const uint8_t *str = (const uint8_t *)s;
-    int i;
+    size_t i;
 
     /* double length, since one octet takes two hex characters */
     length *= 2;
@@ -165,11 +165,11 @@ char *octet_string_hex_string(const void *s, int length)
 static const char b64chars[] = "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
                                "abcdefghijklmnopqrstuvwxyz0123456789+/";
 
-static int base64_block_to_octet_triple(char *out, char *in)
+static size_t base64_block_to_octet_triple(char *out, char *in)
 {
     unsigned char sextets[4] = { 0 };
-    int j = 0;
-    int i;
+    size_t j = 0;
+    size_t i;
 
     for (i = 0; i < 4; i++) {
         char *p = strchr(b64chars, in[i]);
@@ -190,11 +190,11 @@ static int base64_block_to_octet_triple(char *out, char *in)
     return j;
 }
 
-int base64_string_to_octet_string(char *out, int *pad, char *in, int len)
+size_t base64_string_to_octet_string(char *out, int *pad, char *in, size_t len)
 {
-    int k = 0;
-    int i = 0;
-    int j = 0;
+    size_t k = 0;
+    size_t i = 0;
+    size_t j = 0;
 
     if (len % 4 != 0) {
         return 0;
@@ -205,6 +205,6 @@ int base64_string_to_octet_string(char *out, int *pad, char *in, int len)
         k += 3;
         i += 4;
     }
-    *pad = j;
+    *pad = (int)j;
     return i;
 }
